Move tests to old and disable old opmodes
