+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Old/BILDiagonalTest.java
@@ -1,6 +1,8 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Old;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+
+import org.firstinspires.ftc.teamcode.BILAutonomousCommon;
 
 /**
  * Created on 1/10/2017 by Mika.
