+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Old/BILEncoderTest.java
@@ -1,8 +1,10 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Old;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
+
+import org.firstinspires.ftc.teamcode.BILAutonomousCommon;
 
 /**
  * Created by mikab_000 on 1/9/2018.
