+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Old/BILVuforiaBlueBeacons.java
@@ -1,6 +1,7 @@
 package org.firstinspires.ftc.teamcode.Old;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 
 import org.firstinspires.ftc.robotcore.external.navigation.VuforiaLocalizer;
 import org.firstinspires.ftc.robotcore.external.navigation.VuforiaTrackable;
@@ -15,6 +16,7 @@ import java.util.List;
  * Created on 11/12/2016 by Mika.
  */
 @Autonomous(name="BIL: Blue Beacons", group="BIL")
+@Disabled
 public class BILVuforiaBlueBeacons extends BILAutonomousCommon {
     VuforiaLocalizer vuforia;
     BILVuforiaCommon helper = new BILVuforiaCommon();
